refactor: move baseline.ts to serving/lib
